validation

+++ PerfumeStore.API/PerfumeStore.API/Controllers/OrdersController.cs
@@ -1,10 +1,13 @@
 ﻿using AutoMapper;
 using Microsoft.AspNetCore.Mvc;
+using Org.BouncyCastle.Asn1.Ocsp;
 using PerfumeStore.API.Shared.DTO.Request.Order;
+using PerfumeStore.API.Validators;
 using PerfumeStore.Application.Abstractions.Result.Entity;
 using PerfumeStore.Application.Orders;
 using PerfumeStore.Application.Shared.DTO.Request;
 using PerfumeStore.Application.Shared.DTO.Response;
+using PerfumeStore.Domain.Entities.Products;
 
 namespace PerfumeStore.API.Controllers
 {
@@ -14,16 +17,25 @@ namespace PerfumeStore.API.Controllers
     {
         public readonly IOrdersService _orderService;
         private readonly IMapper _mapper;
+        private readonly IValidationService _validationService;
 
-        public OrdersController(IOrdersService orderService, IMapper mapper)
+        public OrdersController(IOrdersService orderService, IMapper mapper, IValidationService validationService)
         {
             _orderService = orderService;
             _mapper = mapper;
+            _validationService = validationService;
         }
 
         [HttpPost]
         public async Task<IActionResult> SubmitOrder([FromBody] CrateOrderDtoApi createOrderRequest)
         {
+            var validationResult = await _validationService.ValidateAsync(createOrderRequest);
+
+            if (!validationResult.IsValid)
+            {
+                return BadRequest(validationResult.Errors);
+            }
+
             CreateOrderDtoApp createOrderDtoApp = _mapper.Map<CreateOrderDtoApp>(createOrderRequest);
 
             EntityResult<OrderResponse> result = await _orderService.CreateOrderAsync(createOrderDtoApp);
@@ -37,6 +49,9 @@ namespace PerfumeStore.API.Controllers
         [HttpGet("{orderId}")]
         public async Task<IActionResult> GetOrderById(int orderId)
         {
+            if (orderId <= 0)
+                return BadRequest("Wrong order id");
+
             EntityResult<OrderResponse> result = await _orderService.GetByIdAsync(orderId);
 
             if (result.IsFailure)
@@ -48,6 +63,9 @@ namespace PerfumeStore.API.Controllers
         [HttpDelete("{orderId}")]
         public async Task<IActionResult> DeleteOrder(int orderId)
         {
+            if (orderId <= 0)
+                return BadRequest("Wrong order id");
+
             EntityResult<OrderResponse> result = await _orderService.DeleteOrderAsync(orderId);
 
             if (result.IsFailure)
@@ -59,6 +77,9 @@ namespace PerfumeStore.API.Controllers
         [HttpPatch("{orderId}/mark-as-deleted")]
         public async Task<IActionResult> MarkOrderAsDeleted(int orderId)
         {
+            if (orderId <= 0)
+                return BadRequest("Wrong order id");
+
             EntityResult<OrderResponse> result = await _orderService.MarkOrderAsDeletedAsync(orderId);
 
             if (result.IsFailure)
@@ -68,7 +89,6 @@ namespace PerfumeStore.API.Controllers
         }
 
         [HttpGet]
-        //[Authorize]
         public async Task<IActionResult> GetOrders()
         {
             EntityResult<IEnumerable<OrdersResDto>> result = await _orderService.GetOrdersAsync();
